requirements.txt
aiofiles==23.1.0
aiogram==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.5
django-filter==23.3
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-api-logger==1.1.14
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
magic-filter==1.0.11
multidict==6.0.4
kombu==5.3.2
packaging==23.1
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
webencodings==0.5.1
yarl==1.9.2
vine==5.0.0
wcwidth==0.2.6
